+++ src/app/dashboard/credit-cards/components/TextFieldUnitNumber.js
@@ -1,0 +1,17 @@
+import { TextField } from '@mui/material'
+import React from 'react'
+import { isIncorrectAmounts } from 'src/utils/missing-value'
+
+export default function TextFieldUnitNumber({item, allocation}) {
+    //console.log(allocation)
+  return (
+    <TextField
+      label={'Units'}
+      value={'0'}
+      disabled={true}
+      variant="outlined"
+      autoComplete="off"
+      error={item?.isSubmitted && isIncorrectAmounts(item)}
+    />
+  )
+}
